clean up WIP implementation of space stations
see #118
see #120

+++ src/Helldivers-2-API/Controllers/V1/SpaceStationController.cs
@@ -13,8 +13,9 @@ public static class SpaceStationController
     /// Fetches a list of all available <see cref="SpaceStation" /> information available.
     /// </summary>
     [ProducesResponseType<List<SpaceStation>>(StatusCodes.Status200OK)]
-    public static async Task<IResult> Index(HttpContext context, IStore<SpaceStation, int> store)
+    public static async Task<IResult> Index(HttpContext context, IStore<SpaceStation, long> store)
     {
+        // TODO: check implementation.
         var stations = await store.AllAsync(context.RequestAborted);
 
         return Results.Ok(stations);
@@ -24,7 +25,7 @@ public static class SpaceStationController
     /// Fetches a specific <see cref="SpaceStation" /> identified by <paramref name="index" />.
     /// </summary>
     [ProducesResponseType<SpaceStation>(StatusCodes.Status200OK)]
-    public static async Task<IResult> Show(HttpContext context, IStore<SpaceStation, int> store, [FromRoute] int index)
+    public static async Task<IResult> Show(HttpContext context, IStore<SpaceStation, long> store, [FromRoute] int index)
     {
         var station = await store.GetAsync(index, context.RequestAborted);
