+++ src/Helldivers-2-Core/Extensions/ServiceCollectionExtensions.cs
@@ -74,7 +74,7 @@ public static class ServiceCollectionExtensions
         services.AddSingleton<IStore<Campaign, int>, CampaignStore>();
         services.AddSingleton<IStore<Models.V1.Assignment, long>, Storage.V1.AssignmentStore>();
         services.AddSingleton<IStore<Dispatch, int>, DispatchStore>();
-        services.AddSingleton<IStore<SpaceStation, int>, SpaceStationStore>();
+        services.AddSingleton<IStore<SpaceStation, long>, SpaceStationStore>();
 
         // Register mappers
         services.AddSingleton<AssignmentMapper>();
