+++ src/Helldivers-2-Core/Storage/V1/SpaceStationStore.cs
@@ -4,8 +4,8 @@ using Helldivers.Models.V1;
 namespace Helldivers.Core.Storage.V1;
 
 /// <inheritdoc cref="IStore{T,TKey}" />
-public sealed class SpaceStationStore : StoreBase<SpaceStation, int>
+public sealed class SpaceStationStore : StoreBase<SpaceStation, long>
 {
     /// <inheritdoc />
-    protected override bool GetAsyncPredicate(SpaceStation station, int index) => station.Id32 == index;
+    protected override bool GetAsyncPredicate(SpaceStation station, long index) => station.Id32 == index;
 }
